@@ -10,18 +10,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.2.1",
+    "@angular/animations": "^15.2.2",
     "@angular/cdk": "^15.2.1",
-    "@angular/common": "^15.0.0",
-    "@angular/compiler": "^15.0.0",
-    "@angular/core": "^15.0.0",
-    "@angular/forms": "^15.0.0",
+    "@angular/common": "^15.2.2",
+    "@angular/compiler": "^15.2.2",
+    "@angular/core": "^15.2.2",
+    "@angular/forms": "^15.2.2",
     "@angular/material": "^15.2.2",
-    "@angular/platform-browser": "^15.0.0",
-    "@angular/platform-browser-dynamic": "^15.0.0",
-    "@angular/router": "^15.0.0",
+    "@angular/platform-browser": "^15.2.2",
+    "@angular/platform-browser-dynamic": "^15.2.2",
+    "@angular/router": "^15.2.2",
     "@auth0/angular-jwt": "^5.1.2",
     "@ng-bootstrap/ng-bootstrap": "^14.0.1",
+    "@sweetalert2/ngx-sweetalert2": "^12.1.0",
     "bootstrap": "^5.2.3",
     "jquery": "^3.6.3",
     "ng-angular-popup": "^0.2.0",
@@ -34,8 +35,8 @@
     "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.0.4",
-    "@angular/cli": "~15.0.4",
+    "@angular-devkit/build-angular": "^15.2.2",
+    "@angular/cli": "~15.2.2",
     "@angular/compiler-cli": "^15.0.0",
     "@types/jasmine": "~4.3.0",
     "jasmine-core": "~4.5.0",
