@@ -30,6 +30,7 @@ export class ApiService {
   // }
 
   deleteCategory(id: number){
+    console.log(id);
 
     const httpOptions = {
       headers: new HttpHeaders({
@@ -39,8 +40,8 @@ export class ApiService {
       })
     };
 
-      return this.http.get(api + `category/delete/${id}}`
-      , {headers:httpOptions.headers, responseType: 'text'})//stringify de chuyen doi tu object sang json
+      return this.http.delete(api + `category/delete/${id}`
+      , {headers:httpOptions.headers, responseType: 'json'})//stringify de chuyen doi tu object sang json
 
 
   }
@@ -56,7 +57,7 @@ export class ApiService {
     };
 
       return this.http.get(api + 'category/list'
-      , {headers:httpOptions.headers, responseType: 'text'})//stringify de chuyen doi tu object sang json
+      , {headers:httpOptions.headers, responseType: 'json'})//stringify de chuyen doi tu object sang json
 
 
   }
@@ -71,7 +72,7 @@ export class ApiService {
     };
     const category = {name: name}
 
-    return this.http.post(api + 'category/add', category, {headers:httpOptions.headers, responseType: 'text'})//stringify de chuyen doi tu object sang json
+    return this.http.post(api + 'category/add', category, {headers:httpOptions.headers, responseType: 'json'})//stringify de chuyen doi tu object sang json
 
   }
 
