@@ -5,11 +5,12 @@ import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { NgToastService } from 'ng-angular-popup';
 import { ApiService } from 'src/app/api.service';
+import Swal from 'sweetalert2'; 
 interface Category {
   id: number,
   name: string,
-  createdAt: string,
-  updatedAt: string
+  // createdAt: string,
+  // updatedAt: string
 
 }
 @Component({
@@ -23,83 +24,125 @@ export class CategorymanagerComponent implements OnInit {
   status: any;
   categoryForm!: FormGroup;
 
-  categories: Category[] = [];
+  // categories: Category[] = [];
+
+  categories: any[] = [];
 
 
   isShowForm = false;
 
   getlistCategory() {
     this.api.getCategory().subscribe((res: any) => {
-      const data = JSON.parse(res)
-      this.categories = data.data;
+      
+     console.log(res.data.list);
+     this.categories = res.data.list
     })
 
 
   }
 
   deleteCategory(id: number) {
-    if (confirm("Are you sure to delete this category?")) {
-      this.api.deleteCategory(id).subscribe(async (res: any) => {
+    
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this imaginary file!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.api.deleteCategory(id).subscribe(async (res: any) => {
 
-        if (res.status == 200) {
-          // await this.getlistCategory();
-          this.router.navigateByUrl('/qam', { skipLocationChange: true }).then(() => {
-            this.router.navigate(['/qam/categorymanager']).then(() => {
-
-              this.toast.success({ detail: "Delete Category Success!", duration: 3000, position: "top-right" })
+          if (res.status == 200) {
+            // await this.getlistCategory();
+            this.router.navigateByUrl('/qam', { skipLocationChange: true }).then(() => {
+              this.router.navigate(['/qam/categorymanager']).then(() => {
+    
+                this.toast.success({ detail: "Delete Category Success!", duration: 3000, position: "top-right" })
+              })
             })
-          })
-
+    
+          }
+    
+        }, (err: any) => {
+          this.toast.error({ detail: "Delete category failed!", duration: 3000, position: "top-right" })
+          console.log(err);
+          // location.reload();
+    
+    
         }
+        )
+        
+      } 
+      // else if (result.dismiss === Swal.DismissReason.cancel) {
+      //   Swal.fire(
+      //     'Cancelled',
+      //     'Your imaginary file is safe :)',
+      //     'error'
+      //   )
+      // }
+    })
+    // if (confirm("Are you sure to delete this category?")) {
+    //   this.api.deleteCategory(id).subscribe(async (res: any) => {
 
-      }, (err: any) => {
-        alert("Delete category failed!");
-        console.log(err);
-        // location.reload();
+    //     if (res.status == 200) {
+    //       // await this.getlistCategory();
+    //       this.router.navigateByUrl('/qam', { skipLocationChange: true }).then(() => {
+    //         this.router.navigate(['/qam/categorymanager']).then(() => {
+
+    //           this.toast.success({ detail: "Delete Category Success!", duration: 3000, position: "top-right" })
+    //         })
+    //       })
+
+    //     }
+
+    //   }, (err: any) => {
+    //     alert("Delete category failed!");
+    //     console.log(err);
+    //     // location.reload();
 
 
-      }
-      )
+    //   }
+    //   )
 
 
-    }
+    // }
 
   }
 
 
   addCategory() {
+    
 
 
-    if (confirm("Are you sure to add this category?")) {
-      
-      
-      this.api.addCategory(this.categoryForm.value.name).subscribe(async (res: any) => {
-        if (res.status == 200) {
-          this.router.navigateByUrl('/qam', { skipLocationChange: true }).then(() => {
-            this.router.navigate(['/qam/categorymanager']).then(() => {
-
-              this.toast.success({ detail: "Add Category Success!", duration: 3000, position: "top-right" })
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this imaginary file!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, add it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.api.addCategory(this.categoryForm.value.name).subscribe(async (res: any) => {
+          if (res.status == 200) {
+            this.router.navigateByUrl('/qam', { skipLocationChange: true }).then(() => {
+              this.router.navigate(['/qam/categorymanager']).then(() => {
+                this.toast.success({ detail: "Add Category Success!", duration: 3000, position: "top-right" })
+              })
             })
-          })
-
-
-
+          }
+        }, (err: any) => {
+          this.toast.error({ detail: "Add category failed!", duration: 3000, position: "top-right" })
+          console.log(err);
+          // location.reload();
         }
-
-      }, (err: any) => {
-        alert("Add category failed!");
-        console.log(err);
-        location.reload();
-
-
+        )
       }
-      )
+    })
 
-
-
-
-    }
-
+    // this.categoryForm.reset();
 
 
   }
@@ -122,7 +165,7 @@ export class CategorymanagerComponent implements OnInit {
       name: new FormControl('', [Validators.required]),
       createdAt: new FormControl(null),
       updatedAt: new FormControl(null)
-    })
+    }, {Validators: false})
 
   } //dependency injection
 
